Fix random `logcat -c` errors
When running jankbench, particularly, occasionally an error of failure to
clear 'mean' or 'system' log is thrown by `logcat -c` causing an
iteration to fail.

Rarer similar errors where seen in exoplayer.

The root cause is believed to be because jankbench spawns a logcat
process to check the run results but never terminates it causing race
conditions in adb logcat logic which results to these errors. Making
sure we terminate the spawned logcat process made tests happy again and
the errors were no longer reproduced.

Note that another form of failure was seen that is believed to be
related is a hang in jankbench when running a test suite with several
workloads and many iterations. The symptom was that jankbench is stuck
in the middle of the test waiting for something to happen. It is
believed that since it depends on logcat to parse its results; it gets
stuck in infinite loop since the result isn't directed to its opened
logcat process it's listening to.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>